+++ src/components/calendar-cell/calendar-cell.ts
@@ -37,6 +37,14 @@ export const calendarCellPropTypes = {
     type: Boolean,
     default: true,
   },
+  subDescription: {
+    type: String,
+    default: '',
+  },
+  icon: {
+    type: String,
+    default: '',
+  },
 };
 
 export type CalendarCellPropTypes = ExtractPropTypes<typeof calendarCellPropTypes>;