+++ src/components/switch/switch.vue
@@ -1,6 +1,6 @@
 <template>
   <div v-bind="switchProps" :class="['spr-flex spr-items-center spr-gap-2', switchTextClass]">
-    <label>
+    <label v-if="!isLeftTextLabel" :for="defaultId" class="spr-cursor-pointer">
       <slot name="leftText">
         <slot></slot>
       </slot>
@@ -16,6 +16,7 @@
         ref="switchRef"
         v-model="proxyValue"
         type="checkbox"
+        :id="defaultId"
         name="checkbox"
         :class="[
           'input spr-absolute spr-left-0 spr-top-0 spr-z-10 spr-m-0 spr-h-6 spr-w-12 spr-opacity-0',
@@ -30,7 +31,7 @@
         ]"
       ></span>
     </div>
-    <label>
+    <label v-if="!isRightTextLabel" :for="defaultId" class="spr-cursor-pointer">
       <slot name="rightText"></slot>
     </label>
   </div>
@@ -45,9 +46,11 @@ import { useSwitch } from './use-switch';
 const props = defineProps(switchPropTypes);
 const emit = defineEmits(switchEmitTypes);
 
+const defaultId = props.id ? props.id + '_' + Math.random().toString(36).substring(2, 8) : 'switch_input_' + Math.random().toString(36).substring(2, 8);
+
 const proxyValue = useVModel(props, 'modelValue', emit);
 
-const { switchWrapperRef, switchRef, switchProps, switchMarkClass, switchTextClass, switchInputClass } =
+const { switchWrapperRef, switchRef, switchProps, switchMarkClass, switchTextClass, switchInputClass, isLeftTextLabel, isRightTextLabel } =
   useSwitch(props);
 </script>