+++ src/components/table/use-table.ts
@@ -10,7 +10,7 @@ export const useTable = (props: TablePropTypes, emit: SetupContext<TableEmitType
     toRefs(props);
   const sortField = ref('');
   const searchField = ref(props.searchModel);
-  const tableSortOrder = ref<TABLE_SORT>(sortOrder.value || 'asc');
+  const tableSortOrder = ref<TABLE_SORT>(sortOrder.value);
   const selectAll = ref(false);
   const selectedData = ref<TableData[]>([]);
 
@@ -25,8 +25,7 @@ export const useTable = (props: TablePropTypes, emit: SetupContext<TableEmitType
   });
 
   const sortedData = computed(() => {
-    if (!sortField.value || sortOrder.value) return dataTable.value;
-
+    if (!sortField.value || sortOrder?.value) return dataTable.value;
     const sorted = [...dataTable.value].sort((a, b) => {
       const fieldA = extractLowerCasedTitle(a[sortField.value]);
       const fieldB = extractLowerCasedTitle(b[sortField.value]);
@@ -47,6 +46,14 @@ export const useTable = (props: TablePropTypes, emit: SetupContext<TableEmitType
     }
 
     emit('onSort', { field: field, sortOrder: tableSortOrder.value });
+  };
+
+  const getSortIcon = (field: string) => {
+    if (sortField.value !== field) {
+      return 'ph:caret-up-down-light';
+    }
+
+    return tableSortOrder.value === 'asc' ? 'ph:arrow-up' : 'ph:arrow-down';
   };
 
   const getHeaderCount = computed(() => (action.value ? headers.value.length + 1 : headers.value.length));
@@ -108,7 +115,7 @@ export const useTable = (props: TablePropTypes, emit: SetupContext<TableEmitType
     const tableHeaderActionsClasses = 'spr-border-color-weak spr-w-full spr-border spr-border-solid';
     const tableCellSlotClasses =
       'spr-background-color-surface spr-text-color-strong spr-font-size-100 spr-font-line-height-100 spr-font-letter-spacing-normal spr-uppercase';
-    const tableRowClasses = classNames('hover:spr-background-color-hover spr-min-h-[60px]', {
+    const tableRowClasses = classNames('spr-min-h-[60px]', {
       'spr-cursor-pointer': props.isRowClickable,
     });
     const tableDataClasses =
@@ -124,7 +131,7 @@ export const useTable = (props: TablePropTypes, emit: SetupContext<TableEmitType
       'spr-h-[360px]': !fullHeight.value && !slots.footer,
     });
 
-    const emptyStateClasses = classNames({
+    const emptyStateBaseClasses = classNames({
       'spr-overflow-y-hidden spr-h-[calc(90vh-150px)] md:spr-h-[calc(80vh-150px)] sm:spr-h-[calc(70vh-150px)]':
         fullHeight.value && slots.footer, // Adjust tbody height for header/footer
       'spr-overflow-y-auto spr-h-[75vh]': fullHeight.value && !slots.footer, // Adjust tbody height for header/footer
@@ -137,6 +144,9 @@ export const useTable = (props: TablePropTypes, emit: SetupContext<TableEmitType
     );
 
     const multiselectClass = classNames('spr-px-size-spacing-2xs spr-py-size-spacing-3xs spr-w-[44px] ');
+
+    const emptyStateClasses = classNames(`${emptyStateBaseClasses} ${props.emptyStateCustomClasses}`);
+    const tableActionSlotClasses = classNames(`${defaultSlotClasses} ${props.tableActionSlotCustomClasses}`);
 
     return {
       headerClasses,
@@ -155,17 +165,18 @@ export const useTable = (props: TablePropTypes, emit: SetupContext<TableEmitType
       tableBodyClasses,
       tableFooterClasses,
       emptyStateClasses,
+      tableActionSlotClasses,
     };
   });
 
-  //assert type TableDataProps 
+  //assert type TableDataProps
   const sortedDataItem = (rowIndex: number, headerField: string) => {
     return sortedData.value[rowIndex][headerField] as TableDataProps;
-  }
+  };
 
   const isTableDataObject = (data: TableDataProps) => {
-    return typeof data === 'object' && 'title' in data
-  }
+    return typeof data === 'object' && 'title' in data;
+  };
 
   const handleSelect = (item: TableData) => {
     if (!selectedKeyId.value || !(selectedKeyId.value in item)) return;
@@ -186,6 +197,8 @@ export const useTable = (props: TablePropTypes, emit: SetupContext<TableEmitType
     } else {
       selectedData.value.push(item);
     }
+
+    emitSelectedData();
   };
 
   const handleSelectAll = () => {
@@ -194,6 +207,8 @@ export const useTable = (props: TablePropTypes, emit: SetupContext<TableEmitType
     } else {
       selectedData.value = [...sortedData.value];
     }
+
+    emitSelectedData();
   };
 
   const isRowSelected = (item: TableData) => {
@@ -210,24 +225,45 @@ export const useTable = (props: TablePropTypes, emit: SetupContext<TableEmitType
     });
   };
 
-  watch(
-    () => selectedData.value.length,
-    () => {
-      if (returnCompleteSelectedProperties.value) {
-        emit('update:selectedData', selectedData.value);
-      } else {
-        const mappedData = selectedData.value.map((item) => {
-          const typedItem = item[selectedKeyId.value] as TableDataProps | string | number;
-          if (typeof typedItem === 'object') {
-            return { ...typedItem };
+  const emitSelectedData = () => {
+    if (returnCompleteSelectedProperties.value) {
+      emit('update:selectedData', selectedData.value);
+    } else {
+      const mappedData = selectedData.value.map((item) => {
+        const typedItem = item[selectedKeyId.value] as TableDataProps | string | number;
+        if (typeof typedItem === 'object') {
+          return { ...typedItem };
+        } else {
+          return typedItem;
+        }
+      });
+      emit('update:selectedData', mappedData);
+    }
+  };
+
+  watch(sortedData, (newVal) => {
+    if (newVal && props.isMultiSelect && selectedData.value.length > 0) {
+      // Remove items from selectedData that are not in the new sortedData
+      // This is to ensure that the selectedData is always in sync with the sortedData
+
+      newVal.forEach((item) => {
+        const selectedIndex = selectedData.value.findIndex((data) => {
+          const typedSelectedData = data[selectedKeyId.value] as TableDataProps;
+          const typedSortedData = item[selectedKeyId.value] as TableDataProps;
+
+          if (isTableDataObject(typedSelectedData) && isTableDataObject(typedSortedData)) {
+            return typedSelectedData.title === typedSortedData.title;
           } else {
-            return typedItem;
+            return data[selectedKeyId.value] === item[selectedKeyId.value];
           }
         });
-        emit('update:selectedData', mappedData);
-      }
-    },
-  );
+
+        if (selectedIndex === -1) {
+          selectedData.value.splice(selectedIndex, 1);
+        }
+      });
+    }
+  });
 
   return {
     sortData,
@@ -246,6 +282,8 @@ export const useTable = (props: TablePropTypes, emit: SetupContext<TableEmitType
     isAllSelected,
     isRowSelected,
     isIndeterminate,
-    sortedDataItem
+    sortedDataItem,
+    getSortIcon,
+    sortField,
   };
 };